// frontend/src/app/generated/patchLeave.ts
/**
 * RESTful API of iHRM
 * The RESTful API of the main core of the iHRM application. Add-ons will have their own API documentation.
 *
 * OpenAPI spec version: 2020.06.14
 * Contact: ihrm@imbus.tn
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface PatchLeave { 
    start?: string;
    isFullDayStart?: boolean;
    end?: string;
    isFullDayEnd?: boolean;
    daysNumber?: number;
    leaveType?: PatchLeave.LeaveTypeEnum;
    description?: string;
}
export namespace PatchLeave {
    export type LeaveTypeEnum = 'Sick' | 'Holiday';
    export const LeaveTypeEnum = {
        Sick: 'Sick' as LeaveTypeEnum,
        Holiday: 'Holiday' as LeaveTypeEnum
    };
}